keep search specs united
